different layouts per auth level is achieved at a high level
