+++ client/src/layouts/Admin.js
@@ -5,7 +5,7 @@ import { Container } from "reactstrap";
 // core components
 import AdminNavbar from "components/Navbars/AdminNavbar.js";
 import AdminFooter from "components/Footers/AdminFooter.js";
-import Sidebar from "components/Sidebar/Sidebar.js";
+import AdminSidebar from "components/Sidebar/AdminSidebar.js";
 
 import routes from "routes.js";
 
@@ -21,7 +21,7 @@ const Admin = (props) => {
 
   const getRoutes = (routes) => {
     return routes.map((prop, key) => {
-      if (prop.layout === "admin" || prop.layout === "general") {
+      if (prop.layout === "admin") {
         return (
           <Route
             path={prop.root + prop.path}
@@ -58,7 +58,7 @@ const Admin = (props) => {
 
   return (
     <>
-      <Sidebar
+      <AdminSidebar
         {...props}
         routes={routes}
         logo={{
