+++ client/src/layouts/Auth.js
@@ -26,10 +26,10 @@ const Auth = (props) => {
 
   const getRoutes = (routes) => {
     return routes.map((prop, key) => {
-      if (prop.layout === "/auth") {
+      if (prop.layout === "auth") {
         return (
           <Route
-            path={prop.layout + prop.path}
+            path={prop.root + prop.path}
             render={() => <prop.component {...props} />}
             key={key}
           />
