+++ client/src/routes.js
@@ -12,7 +12,7 @@ var routes = [
     icon: "ni ni-tv-2 text-primary",
     component: Index,
     layout: "general",
-    root: "/admin",
+    root: "/general",
     display: true,
   },
   {
@@ -21,6 +21,24 @@ var routes = [
     icon: "ni ni-single-copy-04 text-teal",
     component: Tickets,
     layout: "general",
+    root: "/general",
+    display: true,
+  },
+  {
+    path: "/index",
+    name: "Dashboard",
+    icon: "ni ni-tv-2 text-primary",
+    component: Index,
+    layout: "admin",
+    root: "/admin",
+    display: true,
+  },
+  {
+    path: "/tickets",
+    name: "Tickets",
+    icon: "ni ni-single-copy-04 text-teal",
+    component: Tickets,
+    layout: "admin",
     root: "/admin",
     display: true,
   },
@@ -39,7 +57,7 @@ var routes = [
     icon: "ni ni-key-25 text-info",
     component: Login,
     layout: "auth",
-    root: "/admin",
+    root: "/auth",
   },
   {
     path: "/register",
@@ -47,13 +65,21 @@ var routes = [
     icon: "ni ni-circle-08 text-pink",
     component: Register,
     layout: "auth",
-    root: "/admin",
+    root: "/auth",
   },
   {
     path: "/project/:id",
     name: "Project",
     component: Project,
     layout: "general",
+    root: "/general",
+    display: false,
+  },
+  {
+    path: "/project/:id",
+    name: "Project",
+    component: Project,
+    layout: "admin",
     root: "/admin",
     display: false,
   },
